% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/peek.R
\name{composite_tiles}
\alias{composite_tiles}
\title{Stitch a matrix of tile image paths into one magick image}
\usage{
composite_tiles(path_matrix)
}
\description{
Stitch a matrix of tile image paths into one magick image
}
