% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/peek.R
\docType{data}
\name{.peek_state}
\alias{.peek_state}
\title{State: a mutable environment holding the current view}
\format{
An object of class \code{environment} of length 0.
}
\usage{
.peek_state
}
\description{
State: a mutable environment holding the current view
}
\keyword{datasets}
