% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/peek.R
\docType{data}
\name{ESRI_IMAGERY}
\alias{ESRI_IMAGERY}
\title{Default tile URL template (Esri World Imagery)}
\format{
An object of class \code{character} of length 1.
}
\usage{
ESRI_IMAGERY
}
\description{
Default tile URL template (Esri World Imagery)
}
\keyword{datasets}
