% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/peek.R
\name{fetch_tiles}
\alias{fetch_tiles}
\title{Fetch tiles to a cache directory, return file paths as a matrix
rows = y tiles (top to bottom), cols = x tiles (left to right)}
\usage{
fetch_tiles(
  tx_range,
  ty_range,
  zoom,
  template = ESRI_IMAGERY,
  cache_dir = file.path(tempdir(), "peek_cache")
)
}
\description{
Fetch tiles to a cache directory, return file paths as a matrix
rows = y tiles (top to bottom), cols = x tiles (left to right)
}
