% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/peek.R
\docType{data}
\name{ORIGIN}
\alias{ORIGIN}
\title{peek - browse XYZ tiles in the RStudio Viewer via magick}
\format{
An object of class \code{numeric} of length 1.
}
\usage{
ORIGIN
}
\description{
User provides an EPSG:3857 bounding box, peek picks an appropriate zoom,
fetches the tiles, composites them with magick, and displays the result.
Web Mercator full extent
}
\keyword{datasets}
