% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/peek.R
\docType{data}
\name{OSM_TILES}
\alias{OSM_TILES}
\title{OSM as an alternative}
\format{
An object of class \code{character} of length 1.
}
\usage{
OSM_TILES
}
\description{
OSM as an alternative
}
\keyword{datasets}
