% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/peek.R
\name{peek}
\alias{peek}
\title{Browse a region of the world in the RStudio Viewer}
\usage{
peek(bbox, width = 768, template = ESRI_IMAGERY, zmin = 19)
}
\arguments{
\item{bbox}{numeric vector c(xmin, ymin, xmax, ymax) in EPSG:3857}

\item{width}{target pixel width of the composited image}

\item{template}{XYZ tile URL template}
}
\value{
magick image (also displayed in Viewer)
}
\description{
Browse a region of the world in the RStudio Viewer
}
\examples{
# Hobart area (roughly)
# In EPSG:3857:
hobart <- c(16348000, -5310000, 16360000, -5298000)
peek(hobart)
}
