% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/peek.R
\name{pick_zoom}
\alias{pick_zoom}
\title{Choose a zoom level that gives roughly the requested pixel width
for the given EPSG:3857 extent}
\usage{
pick_zoom(xmin, xmax, target_width = 768, zmin = 19)
}
\description{
Choose a zoom level that gives roughly the requested pixel width
for the given EPSG:3857 extent
}
