% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/peek.R
\name{tile_to_xy}
\alias{tile_to_xy}
\title{Tile x/y back to EPSG:3857 origin (top-left corner of that tile)}
\usage{
tile_to_xy(tx, ty, zoom)
}
\description{
Tile x/y back to EPSG:3857 origin (top-left corner of that tile)
}
