% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/peek.R
\name{xy_to_tile}
\alias{xy_to_tile}
\title{Convert EPSG:3857 coordinate to tile index at given zoom
Note: y tile index is flipped (origin top-left in TMS-like XYZ)}
\usage{
xy_to_tile(x, y, zoom)
}
\description{
Convert EPSG:3857 coordinate to tile index at given zoom
Note: y tile index is flipped (origin top-left in TMS-like XYZ)
}
